fix: water entity and id to double

+++ back/migrations/init_waters.sql
@@ -1,5 +1,5 @@
 create table if not exists waters (
-    id serial primary key,
+    id double precision primary key,
     x double precision not null,
     y double precision not null
 )
