+++ back/src/main/java/inno/fishmasters/repository/WaterRepository.java
@@ -3,5 +3,5 @@ package inno.fishmasters.repository;
 import inno.fishmasters.entity.Water;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WaterRepository extends JpaRepository<Water, Long> {
+public interface WaterRepository extends JpaRepository<Water, Double> {
 }
